fix: add (tortured) validation for keywords
(Complicated for many reasons... new items are added to a temp list
that doesn't get merged back to the "real" backing store until Save,
but need to check both the list and the backing store for validation
(e.g. enforcing no duplicates).  So validator needs context from
chip and including parent.  The shadow list immediately has values,
so we can't simply compare, because when the validation rule is
triggered, the value is already in the list.  So, hack, disabled
the + button while one thing is being added until thing is valid,
and validator only checks up to the penultimate element in the temp
list -- e.g. not the one that was just typed.  But to make that work,
we have to disable the + button, and the enclosing form is in the
parent.  So we have to pass the form v-model through to the chip
so we can disable the button when invalid.

Needs a refactor given all the things chips are or are not doing.)